Renaming to KotlinCompiled from KotlinTypesafe
