+++ compiled-parser/build.gradle.kts
