+++ compiled-parser/src/main/kotlin/momosetkn/liquibase/kotlin/parser/KotlinCompiledDatabaseChangeLog.kt
@@ -1,0 +1,5 @@
+package momosetkn.liquibase.kotlin.parser
+
+abstract class KotlinCompiledDatabaseChangeLog(
+    internal val body: momosetkn.liquibase.kotlin.dsl.ChangeLogDsl.() -> Unit
+)
