+++ compiled-parser/src/main/resources/META-INF/services/liquibase.parser.ChangeLogParser
@@ -1,0 +1,1 @@
+momosetkn.liquibase.kotlin.parser.KotlinCompiledLiquibaseChangeLogParser
