+++ compiled-serializer/build.gradle.kts
