+++ compiled-serializer/detekt-baseline.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" ?>
+<SmellBaseline>
+  <ManuallySuppressedIssues></ManuallySuppressedIssues>
+  <CurrentIssues>
+    <ID>InvalidPackageDeclaration:KotlinCompiledChangeLogSerializer.kt$package momosetkn.liquibase.kotlin.serializer</ID>
+  </CurrentIssues>
+</SmellBaseline>
