+++ compiled-serializer/src/main/resources/META-INF/services/liquibase.serializer.ChangeLogSerializer
@@ -1,0 +1,1 @@
+momosetkn.liquibase.kotlin.serializer.KotlinCompiledChangeLogSerializer
