+++ custom-komapper-jdbc-change/src/main/kotlin/momosetkn/liquibase/kotlin/change/CustomKomapperJdbcChangeParserImports.kt
@@ -5,7 +5,7 @@ import momosetkn.liquibase.kotlin.parser.KotlinScriptParserImports
 class CustomKomapperJdbcChangeParserImports : KotlinScriptParserImports {
     override fun imports() = listOf(
         "momosetkn.liquibase.kotlin.change.customKomapperJdbcChange",
-        "momosetkn.liquibase.kotlin.parser.KotlinTypesafeDatabaseChangeLog",
+        "momosetkn.liquibase.kotlin.parser.KotlinCompiledDatabaseChangeLog",
         "org.komapper.core.dsl.QueryDsl",
     )
 }
