+++ integration-test/src/main/kotlin/momosetkn/liquibase/changelogs/CompiledDatabaseChangelogAll.kt
@@ -1,0 +1,8 @@
+package momosetkn.liquibase.changelogs
+
+import momosetkn.liquibase.kotlin.parser.KotlinCompiledDatabaseChangeLog
+
+class CompiledDatabaseChangelogAll : KotlinCompiledDatabaseChangeLog({
+    property(name = "includeDir", value = "momosetkn.liquibase.changelogs.main")
+    includeAll("\${includeDir}")
+})
