--- integration-test/src/main/kotlin/momosetkn/liquibase/changelogs/TypesafeDatabaseChangelogAll.kt
@@ -1,8 +1,0 @@
-package momosetkn.liquibase.changelogs
-
-import momosetkn.liquibase.kotlin.parser.KotlinTypesafeDatabaseChangeLog
-
-class TypesafeDatabaseChangelogAll : KotlinTypesafeDatabaseChangeLog({
-    property(name = "includeDir", value = "momosetkn.liquibase.changelogs.main")
-    includeAll("\${includeDir}")
-})
