+++ integration-test/src/main/resources/KotlinCompiledMigrateAndSerializeSpec/parser_expect/db.ddl-0.sql
