+++ integration-test/src/main/resources/KotlinCompiledMigrateAndSerializeSpec/parser_input/db.changelog-all.kts
@@ -1,0 +1,4 @@
+databaseChangeLog {
+    property(name = "includeDir", value = "KotlinScriptMigrateAndSerializeSpec/parser_input/main/")
+    includeAll("\${includeDir}")
+}
