+++ integration-test/src/main/resources/KotlinCompiledMigrateAndSerializeSpec/parser_input/main/db.changelog-0.kts
@@ -1,0 +1,14 @@
+databaseChangeLog {
+    changeSet(author = "momose (generated)", id = "1715520327312-0") {
+        tagDatabase("started")
+    }
+
+    changeSet(author = "momose (generated)", id = "1715520327312-10") {
+        createTable(tableName = "company") {
+            column(name = "id", type = "UUID") {
+                constraints(nullable = false, primaryKey = true)
+            }
+            column(name = "name", type = "VARCHAR(256)")
+        }
+    }
+}
