+++ integration-test/src/main/resources/KotlinCompiledMigrateAndSerializeSpec/serializer_actual/.gitignore
@@ -1,0 +1,1 @@
+*.kts
