--- integration-test/src/main/resources/KotlinTypesafeMigrateAndSerializeSpec/serializer_actual/.gitignore
@@ -1,1 +1,0 @@
-*.kt
