+++ integration-test/src/main/resources/log4j2.xml
@@ -8,7 +8,7 @@
     </Appenders>
 
     <Loggers>
-        <Root level="info">
+        <Root level="debug">
             <AppenderRef ref="Console"/>
         </Root>
     </Loggers>
