--- typesafe-parser/src/main/kotlin/momosetkn/liquibase/kotlin/parser/KotlinTypesafeDatabaseChangeLog.kt
@@ -1,5 +1,0 @@
-package momosetkn.liquibase.kotlin.parser
-
-abstract class KotlinTypesafeDatabaseChangeLog(
-    internal val body: momosetkn.liquibase.kotlin.dsl.ChangeLogDsl.() -> Unit
-)
