--- typesafe-parser/src/main/resources/META-INF/services/liquibase.parser.ChangeLogParser
@@ -1,1 +1,0 @@
-momosetkn.liquibase.kotlin.parser.KotlinTypesafeLiquibaseChangeLogParser
