--- typesafe-serializer/src/main/resources/META-INF/services/liquibase.serializer.ChangeLogSerializer
@@ -1,1 +1,0 @@
-momosetkn.liquibase.kotlin.serializer.KotlinTypesafeChangeLogSerializer
